requirements.txt
fastapi
uvicorn
python-multipart
ultralytics
opencv-python-headless
easyocr
mysql-connector-python
google-genai
numpy
Pillow
